requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-import-export==3.1.0
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
